requirements.txt
fastapi
uvicorn
google-cloud-speech
requests
bs4
selenium
webdriver-manager
torch
pydantic
google-cloud-aiplatform
google-cloud-secret-manager
groq
python-dotenv
httpx
supabase
python-multipart